+++ R/document.R
@@ -7,22 +7,35 @@
 #' @param pkg package description, can be path or package name.  See
 #'   [as.package()] for more information
 #' @inheritParams roxygen2::roxygenise
+#' @param quiet if `TRUE` suppresses output from this function.
 #' @seealso [roxygen2::roxygenize()],
 #'   `browseVignettes("roxygen2")`
 #' @export
-document <- function(pkg = ".", roclets = NULL) {
+document <- function(pkg = ".", roclets = NULL, quiet = FALSE) {
   pkg <- as.package(pkg)
-  message("Updating ", pkg$package, " documentation")
+  if (!isTRUE(quiet)) {
+    message("Updating ", pkg$package, " documentation")
+  }
 
   save_all()
 
   if (pkg$package == "roxygen2") {
     # roxygen2 crashes if it reloads itself
-    load_all(pkg$path, quiet = TRUE)
+    load_all(pkg$path, quiet = quiet)
     load_code <- function(path) asNamespace("roxygen2")
   } else {
-    load_code <- roxygen2::env_package
+    load_code <- function(path) {
+      pkgload::load_all(path, compile = FALSE, helpers = FALSE, attach_testthat = FALSE, quiet = quiet)$env
+    }
   }
+
+  if (quiet) {
+    output <- tempfile()
+    on.exit(unlink(output))
+
+    withr::local_output_sink(output)
+  }
+
   withr::with_envvar(r_env_vars(),
     roxygen2::roxygenise(pkg$path, roclets, load_code = load_code)
   )